# Avoid copying method names when going through the JSObject method invoke path