--- a/src/mono/wasm/runtime/driver.c
+++ b/src/mono/wasm/runtime/driver.c
@@ -98,32 +98,23 @@ mono_wasm_invoke_js (MonoString *str, int *is_exception)
 	if (str == NULL)
 		return NULL;
 
-	int is_interned = mono_string_instance_is_interned (str);
-	mono_unichar2 *native_chars = mono_string_chars (str);
-	int native_len = mono_string_length (str) * 2;
 	int native_res_len;
 	int *p_native_res_len = &native_res_len;
 
 	mono_unichar2 *native_res = (mono_unichar2*)EM_ASM_INT ({
-		var str;
-		// If the expression is interned, use binding_support's intern table implementation to
-		//  avoid decoding it again unless necessary
-		// We could technically use conv_string for both cases here, but it's more expensive
-		//  than using decode directly in the case where the expression isn't interned
-		if ($4)
-			str = BINDING.conv_string($5, true);
-		else
-			str = MONO.string_decoder.decode ($0, $0 + $1);
+		let str = MONO.string_decoder.copy ($0),
+			p_native_res_len = $1,
+			p_is_exception = $2;
 
 		try {
 			var res = eval (str);
 			if (res === null || res == undefined)
 				return 0;
 			res = res.toString ();
-			setValue ($2, 0, "i32");
+			setValue (p_is_exception, 0, "i32");
 		} catch (e) {
 			res = e.toString();
-			setValue ($2, 1, "i32");
+			setValue (p_is_exception, 1, "i32");
 			if (res === null || res === undefined)
 				res = "unknown exception";
 
@@ -139,9 +130,9 @@ mono_wasm_invoke_js (MonoString *str, int *is_exception)
 		}
 		var buff = Module._malloc((res.length + 1) * 2);
 		stringToUTF16 (res, buff, (res.length + 1) * 2);
-		setValue ($3, res.length, "i32");
+		setValue (p_native_res_len, res.length, "i32");
 		return buff;
-	}, (int)native_chars, native_len, is_exception, p_native_res_len, is_interned, (int)str);
+	}, (int)str, p_native_res_len, is_exception);
 
 	if (native_res == NULL)
 		return NULL;
@@ -731,6 +722,7 @@ mono_wasm_string_get_utf8 (MonoString *str)
 	return mono_string_to_utf8 (str); //XXX JS is responsible for freeing this
 }
 
+// Deprecated
 EMSCRIPTEN_KEEPALIVE void
 mono_wasm_string_convert (MonoString *str)
 {