Right now we always copy the name of the method on every invoke, which is unnecessarily wasteful since it's going to be an interned string literal 99.9% of the time. This PR fixes that, at the cost of a relatively small theoretical performance hit for the case where it's not interned (it will still work, though.)

Benchmark timings before/after:

```
// before
|              InvokeMethodIntViaInvokeJSWithArgs | 17.158 ms | 0.2440 ms | 0.2038 ms | 17.103 ms | 16.893 ms | 17.612 ms |     - |     - |     - | 160,109 B |
| InvokeMethodIntViaInvokeJSWithArgs_InternedName |  9.503 ms | 0.1248 ms | 0.1167 ms |  9.488 ms |  9.345 ms |  9.744 ms |     - |     - |     - | 160,215 B |

// after
|              InvokeMethodIntViaInvokeJSWithArgs | 10.755 ms | 0.0617 ms | 0.0547 ms | 10.748 ms | 10.683 ms | 10.858 ms |     - |     - |     - | 160,104 B |
| InvokeMethodIntViaInvokeJSWithArgs_InternedName | 10.504 ms | 0.0539 ms | 0.0505 ms | 10.507 ms | 10.419 ms | 10.580 ms |     - |     - |     - | 160,216 B |
```

EDIT: I have now reworked this PR to refactor all our string handling code so that library_mono handles the C# -> JS intern table and all string decoding goes through a single path that handles interning, which should deliver good performance for all the bindings APIs and not just InvokeJSWithArgs. This also removes some duplication from the JS -> C# paths.