@kg 2021-06-14
> 
> 
> `mono_wasm_get_object_property`, `mono_wasm_set_object_property`, `mono_wasm_new` and `mono_wasm_get_global_object` could possibly benefit from the same optimization.

Good point, I'll try to generalize this