@kg 2021-06-15
> 
> 
> This makes sense, I do worry that we could end up with a bunch of strings in that table with no visibility and no way to clean them up.

I think the potential downside here isn't too great, since the strings already need to be in the C# intern table for that to happen. So the worst possible scenario is every C# interned string ends up copied into the JS heap, but it's somewhat unlikely. AFAIK the only way that could occur is if they call String.Intern, so they can just Ctrl+F their codebase for that.